+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Support use of nulls in collections when the key or value types are
   explicitly nullable.
 - Allow `JsonObject` to be instantiated from a `Map<dynamic, dynamic>`.
+- Mark nested builder getters in `instantiable: false` classes not nullable,
+  to match the implementations. Use `autoCreateNestedBuilders: false` to get
+  the old behaviour.
 
 # 8.0.6